Implement templating
This change enables the usage of a TEMPLATE_FILENAME environment
variable to specify a template JSON file containing the values of all
the template variables that can be modified for each individual brand's
documentation. All template files are stored under the templates/
directory, to add a slight bit of organization to what could otherwise
be hundreds of files scattered throughout the root directory of this
project. If the TEMPLATE_FILENAME environment variable is not defined,
or if it contains an invalid value, the default value of "NULL" is used
for all template variables.

Solves https://warthogs.atlassian.net/browse/SN-1971