explanation: base-stores: rework into explanation
Improve the explanatory power, specifically targetting the "what" and
"why" of this split between base and device view stores.

What hasn't been covered is the Serial Vault note. This certainly
belongs somewhere, and it needs to be reworked properly as the Model
Service is still as yet unavailable broadly. Both should be documented.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>