reference: configuration-values: heavily rework
Simplify a lot of language, consolidate some notes into the actual text
as there's no reason for them to be notes.

Factored out some support-specific details and will add them to a new
how-to document.

Improved the phrasing and general flow of the document to be a bit more
grokkable, relying on other pages to provide substantial details. This
page effectively acts as a repository for useful information with links
to other, more detailed pages.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>