how-to: configure-serial-vault: reorganize page
Migrate the note to a warning to emphasise its importance.

Provide more succint instructions (which are actually accurate... You
can no longer "click the button in the bottom right" to proceed to the
next doc...) and give the minimum amount of required information.

This page really only exists to isolate the importance of the Serial
Vault.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>