how-to: connecting-devices: move to explanation
Rework the content to make it more useful and consistent. As this
document is really just an explanation for how this whole setup works,
there's no reason to make this a how-to. There may be good reason
to provide a how-to in the future when true and proper secure device
onboarding exists.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>